+++ svelte/types/global.d.ts
@@ -1,0 +1,7 @@
+import { APICallRecord } from './index';
+
+declare global {
+  interface Window {
+    __apiCallRecord?: APICallRecord = {};
+  }
+};